refactor compile system and make the apiviewer work
