+++ Manifest.json
@@ -29,12 +29,12 @@
   "requires": {
     "@qooxdoo/framework": "^6.0.0-beta",
     "@qooxdoo/compiler": "^1.0.0-beta.20200314-0952",
-    "qooxdoo/qxl.apiviewer": "^1.0.0-beta.7",
     "ITISFoundation/qx-osparc-theme": "^0.4.13",
     "johnspackman/UploadMgr": "^1.0.4",
     "cboulanger/qx-contrib-Dialog": "^2.1.1",
     "qooxdoo/qxl.versionlabel": "^1.0.0-beta.3",
-    "ITISFoundation/qx-iconfont-material": "^0.1.4"
+    "ITISFoundation/qx-iconfont-material": "^0.1.4",
+    "qooxdoo/qxl.apiviewer": "^1.0.0-beta.7"
   },
   "externalResources": {
     "script": [
