+++ package.json
@@ -6,7 +6,14 @@
     "type": "git",
     "url": "https://github.com/cometvisu/cometvisu.git"
   },
-  "scripts": {},
+  "scripts": {
+    "presource": "mkdir -p compiled/source && php -S 0.0.0.0:8082 -t compiled/source &",
+    "source": "qx compile --watch",
+    "prebuild": "mkdir -p compiled/build && php -S 0.0.0.0:8082 -t compiled/build &",
+    "build": "qx compile --watch --t build --set fakeLogin=true",
+    "make-cv": "qx compile --t build",
+    "compile": "qx compile --watch"
+  },
   "license": "GPL-3.0",
   "authors": [
     {
