+++ qx-lock.json
@@ -1,14 +1,6 @@
 {
   "libraries": [
     {
-      "library_name": "Qooxdoo API Viewer",
-      "library_version": "1.0.0-beta.7",
-      "path": "qx_packages/qooxdoo_qxl_apiviewer_v1_0_0-beta_7",
-      "uri": "qooxdoo/qxl.apiviewer",
-      "repo_name": "qooxdoo/qxl.apiviewer",
-      "repo_tag": "v1.0.0-beta.7"
-    },
-    {
       "library_name": "versionlabel",
       "library_version": "1.0.0-beta.3",
       "path": "qx_packages/qooxdoo_qxl_versionlabel_v1_0_0-beta_3",
@@ -47,6 +39,14 @@
       "uri": "cboulanger/qx-contrib-Dialog",
       "repo_name": "cboulanger/qx-contrib-Dialog",
       "repo_tag": "v2.1.1"
+    },
+    {
+      "library_name": "Qooxdoo API Viewer",
+      "library_version": "1.0.0-beta.7",
+      "path": "qx_packages/qooxdoo_qxl_apiviewer_v1_0_0-beta_7",
+      "uri": "qooxdoo/qxl.apiviewer",
+      "repo_name": "qooxdoo/qxl.apiviewer",
+      "repo_tag": "v1.0.0-beta.7"
     }
   ],
   "version": "2.1.0"
